feat: move large mixin methods to MixinHelpers
