+++ src/main/java/com/gregtechceu/gtceu/common/data/GTFluids.java
@@ -1,9 +1,12 @@
 package com.gregtechceu.gtceu.common.data;
 
+import com.gregtechceu.gtceu.api.GTCEuAPI;
 import com.gregtechceu.gtceu.api.data.chemical.material.Material;
 import com.gregtechceu.gtceu.api.data.chemical.material.properties.PropertyKey;
+import com.gregtechceu.gtceu.api.data.chemical.material.registry.MaterialRegistry;
 import com.gregtechceu.gtceu.api.fluids.store.FluidStorageKeys;
 import com.gregtechceu.gtceu.api.registry.GTRegistries;
+import com.gregtechceu.gtceu.api.registry.registrate.GTRegistrate;
 import net.minecraft.world.level.material.Fluid;
 import net.minecraft.world.level.material.Fluids;
 
@@ -24,11 +27,14 @@ public class GTFluids {
         handleNonMaterialFluids(GTMaterials.Lava, Fluids.LAVA);
         REGISTRATE.creativeModeTab(() -> GTCreativeModeTabs.MATERIAL_FLUID);
         // register fluids for materials
-        for (var material : GTRegistries.MATERIALS) {
-            var fluidProperty = material.getProperty(PropertyKey.FLUID);
+        for (MaterialRegistry registry : GTCEuAPI.materialManager.getRegistries()) {
+            GTRegistrate registrate = registry.getRegistrate();
+            for (var material : registry.getAllMaterials()) {
+                var fluidProperty = material.getProperty(PropertyKey.FLUID);
 
-            if (fluidProperty != null) {
-                fluidProperty.getStorage().registerFluids(material);
+                if (fluidProperty != null) {
+                    fluidProperty.getStorage().registerFluids(material, registrate);
+                }
             }
         }
     }
