+++ src/main/java/com/gregtechceu/gtceu/core/mixins/LootDataManagerMixin.java
@@ -1,36 +1,17 @@
 package com.gregtechceu.gtceu.core.mixins;
 
-import com.gregtechceu.gtceu.api.data.chemical.ChemicalHelper;
-import com.gregtechceu.gtceu.api.data.chemical.material.Material;
-import com.gregtechceu.gtceu.api.data.chemical.material.properties.PropertyKey;
-import com.gregtechceu.gtceu.api.data.tag.TagPrefix;
-import com.gregtechceu.gtceu.api.registry.GTRegistries;
-import com.gregtechceu.gtceu.common.data.GTBlocks;
-import com.gregtechceu.gtceu.config.ConfigHolder;
 import com.gregtechceu.gtceu.core.MixinHelpers;
-import com.gregtechceu.gtceu.data.recipe.CustomTags;
 import com.lowdragmc.lowdraglib.Platform;
-import net.minecraft.data.loot.BlockLootSubProvider;
-import net.minecraft.data.loot.packs.VanillaBlockLoot;
 import net.minecraft.resources.ResourceLocation;
-import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.enchantment.Enchantments;
-import net.minecraft.world.level.block.Block;
-import net.minecraft.world.level.storage.loot.*;
-import net.minecraft.world.level.storage.loot.entries.LootItem;
-import net.minecraft.world.level.storage.loot.functions.ApplyBonusCount;
-import net.minecraft.world.level.storage.loot.functions.ApplyExplosionDecay;
-import net.minecraft.world.level.storage.loot.functions.LimitCount;
-import net.minecraft.world.level.storage.loot.functions.SetItemCountFunction;
-import net.minecraft.world.level.storage.loot.parameters.LootContextParamSets;
-import net.minecraft.world.level.storage.loot.providers.number.UniformGenerator;
+import net.minecraft.world.level.storage.loot.LootDataManager;
+import net.minecraft.world.level.storage.loot.LootDataType;
+import net.minecraft.world.level.storage.loot.LootTable;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 import java.util.Map;
-import java.util.function.Supplier;
 
 @Mixin(LootDataManager.class)
 public abstract class LootDataManagerMixin {
@@ -41,68 +22,6 @@ public abstract class LootDataManagerMixin {
         if (Platform.isDatagen()) return;
 
         Map<ResourceLocation, LootTable> lootTables = (Map<ResourceLocation, LootTable>) allElements.get(LootDataType.TABLE);
-
-        GTBlocks.MATERIAL_BLOCKS.rowMap().forEach((prefix, map) -> {
-            if (TagPrefix.ORES.containsKey(prefix)) {
-                map.forEach((material, blockEntry) -> {
-                    ResourceLocation lootTableId = new ResourceLocation(blockEntry.getId().getNamespace(), "blocks/" + blockEntry.getId().getPath());
-                    Block block = blockEntry.get();
-
-                    if ((prefix != TagPrefix.oreNetherrack && prefix != TagPrefix.oreEndstone) && !ConfigHolder.INSTANCE.worldgen.allUniqueStoneTypes) {
-                        TagPrefix orePrefix = TagPrefix.ORES.get(prefix).isNether() ? TagPrefix.ORES.get(prefix).stoneType().get().is(CustomTags.ENDSTONE_ORE_REPLACEABLES) ? TagPrefix.oreEndstone : TagPrefix.oreNetherrack : TagPrefix.ore;
-                        block = ChemicalHelper.getBlock(orePrefix, material);
-                    }
-
-                    ItemStack dropItem = ChemicalHelper.get(TagPrefix.rawOre, material);
-                    if (dropItem.isEmpty()) dropItem = ChemicalHelper.get(TagPrefix.gem, material);
-                    if (dropItem.isEmpty()) dropItem = ChemicalHelper.get(TagPrefix.dust, material);
-                    int oreMultiplier = TagPrefix.ORES.get(prefix).isNether() ? 2 : 1;
-
-                    LootTable.Builder builder = BlockLootSubProvider.createSilkTouchDispatchTable(block,
-                            new VanillaBlockLoot().applyExplosionDecay(block,
-                                    LootItem.lootTableItem(dropItem.getItem())
-                                            .apply(SetItemCountFunction.setCount(UniformGenerator.between(1, Math.max(1, material.getProperty(PropertyKey.ORE).getOreMultiplier() * oreMultiplier))))));
-                                            //.apply(ApplyBonusCount.addOreBonusCount(Enchantments.BLOCK_FORTUNE)))); //disable fortune for balance reasons. (for now, until we can think of a better solution.)
-
-                    Supplier<Material> outputDustMat = TagPrefix.ORES.get(prefix).material();
-                    if (outputDustMat != null) {
-                        builder.withPool(LootPool.lootPool().add(
-                                LootItem.lootTableItem(ChemicalHelper.get(TagPrefix.dust, outputDustMat.get()).getItem())
-                                    .when(BlockLootSubProvider.HAS_NO_SILK_TOUCH)
-                                    .apply(SetItemCountFunction.setCount(UniformGenerator.between(0, 1)))
-                                    .apply(ApplyBonusCount.addUniformBonusCount(Enchantments.BLOCK_FORTUNE))
-                                    .apply(LimitCount.limitCount(IntRange.range(0, 2)))
-                                    .apply(ApplyExplosionDecay.explosionDecay())));
-                    }
-                    lootTables.put(lootTableId, builder.setParamSet(LootContextParamSets.BLOCK).build());
-                    ((BlockBehaviourAccessor)blockEntry.get()).setDrops(lootTableId);
-                });
-            } else {
-                MixinHelpers.addMaterialBlockLootTables(lootTables, prefix, map);
-            }
-        });
-        GTBlocks.CABLE_BLOCKS.rowMap().forEach((prefix, map) -> {
-            MixinHelpers.addMaterialBlockLootTables(lootTables, prefix, map);
-        });
-        GTBlocks.FLUID_PIPE_BLOCKS.rowMap().forEach((prefix, map) -> {
-            MixinHelpers.addMaterialBlockLootTables(lootTables, prefix, map);
-        });
-        GTBlocks.ITEM_PIPE_BLOCKS.rowMap().forEach((prefix, map) -> {
-            MixinHelpers.addMaterialBlockLootTables(lootTables, prefix, map);
-        });
-        GTBlocks.SURFACE_ROCK_BLOCKS.forEach((material, blockEntry) -> {
-            ResourceLocation lootTableId = new ResourceLocation(blockEntry.getId().getNamespace(), "blocks/" + blockEntry.getId().getPath());
-            LootTable.Builder builder = new VanillaBlockLoot().createSingleItemTable(ChemicalHelper.get(TagPrefix.dustTiny, material).getItem(), UniformGenerator.between(3, 5))
-                    .apply(ApplyBonusCount.addUniformBonusCount(Enchantments.BLOCK_FORTUNE));
-            lootTables.put(lootTableId, builder.setParamSet(LootContextParamSets.BLOCK).build());
-            ((BlockBehaviourAccessor) blockEntry.get()).setDrops(lootTableId);
-        });
-        GTRegistries.MACHINES.forEach(machine -> {
-            Block block = machine.getBlock();
-            ResourceLocation id = machine.getId();
-            ResourceLocation lootTableId = new ResourceLocation(id.getNamespace(), "blocks/" + id.getPath());
-            ((BlockBehaviourAccessor)block).setDrops(lootTableId);
-            lootTables.put(lootTableId, new VanillaBlockLoot().createSingleItemTable(block).setParamSet(LootContextParamSets.BLOCK).build());
-        });
+        MixinHelpers.generateGTDynamicLoot(lootTables);
     }
 }
